Settings Page
Toggle switches in the settings page to enable/disable the CPU/GPU miners. Changes are saved upon press of the "EXIT" button, and pressing play on the main page will accurately launch the corresponding miner.

Rather ugly looking for now, both in terms of visuals and of the actual code, but it's a functioning first draft.